refactor: [order] 공통모듈 변경사항 반영 리팩토링
