+++ order/build.gradle
@@ -43,8 +43,8 @@ dependencies {
 	annotationProcessor 'org.projectlombok:lombok'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-    implementation 'com.github.sharedeffort:common-module:v1.1.8'
-    implementation 'com.github.sharedeffort:common-jpa:v1.0.1'
+    implementation 'com.github.sharedeffort:common-module:v1.1.9'
+    implementation 'com.github.sharedeffort:common-jpa:v1.0.2'
     implementation 'org.springframework.boot:spring-boot-starter-security'
 
 }
