+++ order/src/main/java/com/high/order/OrderApplication.java
@@ -2,8 +2,10 @@ package com.high.order;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackages = {"com.high.order", "com.library.module"})
 public class OrderApplication {
 
 	public static void main(String[] args) {
