+++ order/src/main/java/com/high/order/application/exception/OrderNotFoundException.java
@@ -3,9 +3,10 @@ package com.high.order.application.exception;
 import com.high.order.exception.OrderErrorCode;
 import com.library.module.exception.CustomException;
 
-public class OrderBadRequestException extends CustomException {
+public class OrderNotFoundException extends CustomException {
 
-    public OrderBadRequestException() {
-        super(OrderErrorCode.BAD_REQUEST);
+
+    public OrderNotFoundException() {
+        super(OrderErrorCode.ORDER_NOT_FOUND);
     }
 }
