+++ order/src/main/java/com/high/order/application/service/OrderService.java
@@ -11,7 +11,7 @@ import com.high.order.application.dto.response.OrderListResponse;
 import com.high.order.application.dto.response.OrderResponse;
 import com.high.order.application.exception.OrderBadRequestException;
 import com.high.order.domain.exception.OrderItemNotFoundExeption;
-import com.high.order.domain.exception.OrderNotFoundException;
+import com.high.order.application.exception.OrderNotFoundException;
 import com.high.order.domain.entity.Order;
 import com.high.order.domain.entity.OrderItem;
 import com.high.order.domain.repository.OrderItemRepository;
@@ -120,7 +120,7 @@ public class OrderService {
             throw new OrderNotFoundException();
         }
         //삭제자 임시
-        order.softDelete(UUID.randomUUID().toString());
+        order.softDelete(UUID.randomUUID());
     }
 
     @Transactional
