+++ order/src/main/java/com/high/order/domain/exception/InvalidOrderStateException.java
@@ -1,13 +1,11 @@
 package com.high.order.domain.exception;
 
+import com.high.order.exception.OrderErrorCode;
 import com.library.module.exception.CustomException;
-import org.springframework.http.HttpStatus;
 
 public class InvalidOrderStateException extends CustomException {
 
-    private static final int INVALID_ORDER_STATUS = 4003;
-
     public InvalidOrderStateException() {
-        super(HttpStatus.BAD_REQUEST, INVALID_ORDER_STATUS, "주문 상태를 변경할 수 없습니다.");
+        super(OrderErrorCode.INVALID_ORDER_STATUS);
     }
 }
