+++ order/src/main/java/com/high/order/domain/exception/OrderCancellationException.java
@@ -1,14 +1,11 @@
 package com.high.order.domain.exception;
 
+import com.high.order.exception.OrderErrorCode;
 import com.library.module.exception.CustomException;
-import org.springframework.http.HttpStatus;
 
 public class OrderCancellationException extends CustomException {
 
-    private static final int ORDER_CANCELLATION_NOT_ALLOWED = 4004;
-
-
     public OrderCancellationException() {
-        super(HttpStatus.BAD_REQUEST, ORDER_CANCELLATION_NOT_ALLOWED, "주문을 취소할 수 없습니다.");
+        super(OrderErrorCode.ORDER_CANCELLATION_NOT_ALLOWED);
     }
 }
