--- order/src/main/java/com/high/order/infrastructure/config/CustomExceptionHandler.java
@@ -1,22 +1,0 @@
-package com.high.order.infrastructure.config;
-
-import com.library.module.common.handler.GlobalExceptionHandler;
-import java.util.HashMap;
-import java.util.Map;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class CustomExceptionHandler extends GlobalExceptionHandler {
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<Map<String, String>> handleValidationExceptions(
-        MethodArgumentNotValidException ex) {
-        Map<String, String> errors = new HashMap<>();
-        ex.getBindingResult().getFieldErrors().forEach(error -> {
-            errors.put(error.getField(), error.getDefaultMessage()); // 여기서 message 가져옴
-        });
-        return ResponseEntity.badRequest().body(errors);
-    }
-}
