+++ order/src/main/java/com/high/order/infrastructure/config/JpaAuditingConfig.java
@@ -1,6 +1,7 @@
 package com.high.order.infrastructure.config;
 
 import java.util.Optional;
+import java.util.UUID;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
@@ -13,15 +14,15 @@ import org.springframework.security.core.context.SecurityContextHolder;
 public class JpaAuditingConfig {
 
     @Bean
-    public AuditorAware<String> auditorProvider() {
+    public AuditorAware<UUID> auditorProvider() {
         return () -> {
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
             if (authentication == null || !authentication.isAuthenticated()) {
-                return Optional.of("SYSTEM");
+                return Optional.of(UUID.randomUUID());
             }
 
-            return Optional.of(authentication.getName());
+            return Optional.of(UUID.randomUUID());
         };
     }
 }
